requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
python-dotenv==0.20.0
psycopg2==2.9.3
Flask-Migrate==3.1.0